pyproject.toml
[tool.poetry]
name = "mailing-service-django"
version = "0.1.0"
description = ""
authors = ["Anton Gromov <seed14112@yaandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2.1"
python-dotenv = "^1.1.0"
psycopg2-binary = "^2.9.10"
djangorestframework = "^3.16.0"
redis = "^6.1.0"
celery = "^5.5.2"
requests = "^2.32.3"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.2.0"
black = "^25.1.0"
isort = "^6.0.1"
pre-commit = "^4.2.0"
flake8-docstrings = "^1.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 119
extend-exclude = "migrations/"

[tool.isort]
profile = "black"
include_trailing_comma = true
line_length = 119
lines_after_imports = 2
multi_line_output = 3
skip = ["__init__.py", "/migrations"]

